--- src/main/java/com/ea/inzynierka/exception/GlobalExceptionHandler.java
@@ -1,40 +1,0 @@
-package com.ea.inzynierka.exception;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.core.annotation.AnnotationUtils;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.servlet.ModelAndView;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-import javax.servlet.http.HttpServletRequest;
-
-@ControllerAdvice
-public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
-	public static final String DEFAULT_ERROR_VIEW = "error";
-
-	private final Logger logger = LoggerFactory.getLogger(GlobalExceptionHandler.class);
-	
-	@ExceptionHandler(value = Exception.class)
-	public ModelAndView defaultErrorHandler(HttpServletRequest req, Exception e) throws Exception {
-		
-		logger.error("[URL] : {}", req.getRequestURL(), e);
-		
-		// If the exception is annotated with @ResponseStatus rethrow it and let
-		// the framework handle it - like the OrderNotFoundException example
-		// at the start of this post.
-		// AnnotationUtils is a Spring Framework utility class.
-		if (AnnotationUtils.findAnnotation(e.getClass(), ResponseStatus.class) != null)
-			throw e;
-
-		// Otherwise setup and send the user to a default error-view.
-		ModelAndView mav = new ModelAndView();
-		mav.addObject("exception", e);
-		mav.addObject("url", req.getRequestURL());
-		mav.setViewName(DEFAULT_ERROR_VIEW);
-		return mav;
-	}
-	
-}
