+++ src/main/java/com/ea/inzynierka/service/BookService.java
@@ -3,7 +3,6 @@ package com.ea.inzynierka.service;
 import com.ea.inzynierka.exception.BookNotFound;
 import com.ea.inzynierka.model.Book;
 import java.util.List;
-import java.util.Optional;
 
 public interface BookService {
 
@@ -11,5 +10,6 @@ public interface BookService {
     Book edit(Book book) throws BookNotFound;
     Book delete(long id) throws BookNotFound;
     List<Book> findAll();
+    Book findById(long id);
 
 }
