--- src/main/java/com/ea/inzynierka/service/UserService.java
@@ -1,44 +1,0 @@
-/*
-package com.gpch.login.service;
-
-import com.gpch.login.model.Role;
-import com.gpch.login.model.User;
-import com.gpch.login.repo.RoleRepository;
-import com.gpch.login.repo.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Service;
-
-import java.util.Arrays;
-import java.util.HashSet;
-
-@Service("userService")
-public class UserService {
-
-    private UserRepository userRepository;
-    private RoleRepository roleRepository;
-    private BCryptPasswordEncoder bCryptPasswordEncoder;
-
-    @Autowired
-    public UserService(@Qualifier("userRepository") UserRepository userRepository,
-                       @Qualifier("roleRepository") RoleRepository roleRepository,
-                       BCryptPasswordEncoder bCryptPasswordEncoder) {
-        this.userRepository = userRepository;
-        this.roleRepository = roleRepository;
-        this.bCryptPasswordEncoder = bCryptPasswordEncoder;
-    }
-
-    public User findUserByEmail(String email) {
-        return userRepository.findByEmail(email);
-    }
-
-    public User saveUser(User user) {
-        user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
-        user.setActive(1);
-        Role userRole = roleRepository.findByRole("ADMIN");
-        user.setRoles(new HashSet<Role>(Arrays.asList(userRole)));
-        return userRepository.save(user);
-    }
-
-}*/
