--- src/main/resources/static/css/home.css
@@ -1,5 +1,0 @@
-.admin-message-text {
-    font-style: normal;
-    font-size: 22px;
-    color: #004080;
-}
