--- src/main/resources/static/css/info.css
@@ -1,5 +1,0 @@
-.validation-message {
-    font-style: normal;
-    font-size: 12px;
-    color: #FF1C19;
-}
