--- src/main/resources/static/css/login.css
@@ -1,36 +1,0 @@
-.wrapper {
-    margin-top: 80px;
-    margin-bottom: 20px;
-}
-
-.form-signin {
-    max-width: 420px;
-    padding: 30px 38px 66px;
-    margin: 0 auto;
-    background-color: #eee;
-    border: 3px dotted rgba(0,0,0,0.1);
-}
-
-.form-signin-heading {
-    text-align:center;
-    margin-bottom: 30px;
-}
-
-.form-control {
-    position: relative;
-    font-size: 16px;
-    height: auto;
-    padding: 10px;
-}
-
-input[type="text"] {
-    margin-bottom: 0px;
-    border-bottom-left-radius: 0;
-    border-bottom-right-radius: 0;
-}
-
-input[type="password"] {
-    margin-bottom: 20px;
-    border-top-left-radius: 0;
-    border-top-right-radius: 0;
-}
